salle_reservation/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
pydantic
bcrypt
aiohttp
requests
psycopg2-binary
httpx


